[Service, Test] 서비스 기능중 수정, 삭제 기능에 예외 처리 추가 및 테스트 코드 수정

+++ src/main/java/threads/server/domain/follow/FollowService.java
@@ -3,8 +3,10 @@ package threads.server.domain.follow;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import threads.server.application.exception.NotFoundException;
+import threads.server.application.exception.UnauthorizedException;
 import threads.server.domain.follow.dto.FollowDTO;
 import threads.server.domain.follow.dto.FollowingDTO;
+import threads.server.domain.follow.dto.UnfollowingDTO;
 import threads.server.domain.user.User;
 
 import java.time.LocalDateTime;
@@ -14,14 +16,18 @@ import java.time.LocalDateTime;
 public class FollowService {
     private final FollowRepository followRepository;
 
-    public void follow(FollowingDTO followDTO) {
-        User toUser = new User(followDTO.toUserId());
-        User fromUser = new User(followDTO.fromUserId());
+    public void follow(FollowingDTO followDto) {
+        User toUser = new User(followDto.toUserId());
+        User fromUser = new User(followDto.fromUserId());
         followRepository.save(new Follow(null, toUser, fromUser, LocalDateTime.now()));
     }
 
-    public void unfollow(Long followId) {
-        Follow follow = followRepository.findById(followId).orElseThrow(() -> new NotFoundException("팔로우 내역을 찾을 수 없습니다."));
+    public void unfollow(UnfollowingDTO followDto) {
+        Follow follow = followRepository.findById(followDto.id()).orElseThrow(() -> new NotFoundException("팔로우 내역을 찾을 수 없습니다."));
+        if(!followDto.fromUserId().equals(follow.getFromUser().getId()) ||
+            !followDto.toUserId().equals(follow.getToUser().getId())) {
+            throw new UnauthorizedException("권한이 없습니다.");
+        }
         followRepository.delete(follow);
     }
 }
